requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1 